requirements.txt
azure-ai-formrecognizer==3.2.1
azure-storage-blob==12.8.1
azure-common==1.1.28
azure-core==1.27.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
filelock==3.12.2
idna==3.4
isodate==0.6.1
msrest==0.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.7.0
urllib3==2.0.3
virtualenv==20.23.1
